Add build flag in Makefile to set MAX_TPM_CONTEXT_COUNT
A TPM_CONTEXT represents a vTPM intance. Currently we use an array to host
the TPM_CONTEXT. Its size is fixed (MAX_TPM_CONTEXT_COUNT). The benefit
of using a fixed size array is that it will never fail because of
out_of_memory in run-time. But it is not flexible that the array size
has to be changed in source code.

This patch adds a build flag to set MAX_TPM_CONTEXT_COUNT. Its default
value is 256.

Signed-off-by: Min Xu <min.m.xu@intel.com>