Add platParameter in _plat__TPM_Initialize as input parameter
Fix #18

In the previous implementation NVM is backed up in a file. When a TPM
instance is initialized, the NVM is restored from a file. But file
system is not supported in vtpm-td. So the NVM is backed up in memory
which is managed by vtpm-td. When TPM instance is initialized, the NVM
content is sent back to the TPM instance so that NVM can be restored.

This patch adds a new input parameter (platParameter) in
_plat__TPM_Initialize so that it can then passed to _plat__NVEnable.

Signed-off-by: Min Xu <min.m.xu@intel.com>